优化：章节id转雪花id

+++ Sourcecode/Song.Entities/Course.cs
@@ -38,7 +38,7 @@ namespace Song.Entities {
     		
     		protected Int32 _Cou_ViewNum;
     		
-    		protected Int32 _Sbj_ID;
+    		protected Int64 _Sbj_ID;
     		
     		protected String _Sbj_Name;
     		
@@ -236,7 +236,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Sbj_ID {
+    		public Int64 Sbj_ID {
     			get {
     				return this._Sbj_ID;
     			}
@@ -633,7 +633,7 @@ namespace Song.Entities {
     				this._Cou_ViewNum = reader.GetInt32(_.Cou_ViewNum);
     			}
     			if ((false == reader.IsDBNull(_.Sbj_ID))) {
-    				this._Sbj_ID = reader.GetInt32(_.Sbj_ID);
+    				this._Sbj_ID = reader.GetInt64(_.Sbj_ID);
     			}
     			if ((false == reader.IsDBNull(_.Sbj_Name))) {
     				this._Sbj_Name = reader.GetString(_.Sbj_Name);
@@ -806,7 +806,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Cou_ViewNum = new WeiSha.Data.Field<Course>("Cou_ViewNum");
     			
     			/// <summary>
-    			/// 字段名：Sbj_ID - 数据类型：Int32
+    			/// 字段名：Sbj_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Sbj_ID = new WeiSha.Data.Field<Course>("Sbj_ID");
     			
@@ -927,3 +927,4 @@ namespace Song.Entities {
     		}
     	}
     }
+
