+++ Sourcecode/Song.Entities/ExamResults.cs
@@ -34,7 +34,7 @@ namespace Song.Entities {
     		
     		protected String _Exam_Name;
     		
-    		protected Int32 _Sbj_ID;
+    		protected Int64 _Sbj_ID;
     		
     		protected String _Sbj_Name;
     		
@@ -242,7 +242,7 @@ namespace Song.Entities {
     		/// <summary>
     		/// -1
     		/// </summary>
-    		public Int32 Sbj_ID {
+    		public Int64 Sbj_ID {
     			get {
     				return this._Sbj_ID;
     			}
@@ -582,7 +582,7 @@ namespace Song.Entities {
     				this._Exam_Name = reader.GetString(_.Exam_Name);
     			}
     			if ((false == reader.IsDBNull(_.Sbj_ID))) {
-    				this._Sbj_ID = reader.GetInt32(_.Sbj_ID);
+    				this._Sbj_ID = reader.GetInt64(_.Sbj_ID);
     			}
     			if ((false == reader.IsDBNull(_.Sbj_Name))) {
     				this._Sbj_Name = reader.GetString(_.Sbj_Name);
@@ -727,7 +727,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Exam_Name = new WeiSha.Data.Field<ExamResults>("Exam_Name");
     			
     			/// <summary>
-    			/// -1 - 字段名：Sbj_ID - 数据类型：Int32
+    			/// -1 - 字段名：Sbj_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Sbj_ID = new WeiSha.Data.Field<ExamResults>("Sbj_ID");
