+++ Sourcecode/Song.Entities/Examination.cs
@@ -28,7 +28,7 @@ namespace Song.Entities {
     		
     		protected Int32 _Exam_Span;
     		
-    		protected Int32 _Sbj_ID;
+    		protected Int64 _Sbj_ID;
     		
     		protected String _Sbj_Name;
     		
@@ -195,7 +195,7 @@ namespace Song.Entities {
     		/// <summary>
     		/// -1
     		/// </summary>
-    		public Int32 Sbj_ID {
+    		public Int64 Sbj_ID {
     			get {
     				return this._Sbj_ID;
     			}
@@ -508,7 +508,7 @@ namespace Song.Entities {
     				this._Exam_Span = reader.GetInt32(_.Exam_Span);
     			}
     			if ((false == reader.IsDBNull(_.Sbj_ID))) {
-    				this._Sbj_ID = reader.GetInt32(_.Sbj_ID);
+    				this._Sbj_ID = reader.GetInt64(_.Sbj_ID);
     			}
     			if ((false == reader.IsDBNull(_.Sbj_Name))) {
     				this._Sbj_Name = reader.GetString(_.Sbj_Name);
@@ -635,7 +635,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Exam_Span = new WeiSha.Data.Field<Examination>("Exam_Span");
     			
     			/// <summary>
-    			/// -1 - 字段名：Sbj_ID - 数据类型：Int32
+    			/// -1 - 字段名：Sbj_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Sbj_ID = new WeiSha.Data.Field<Examination>("Sbj_ID");
