+++ Sourcecode/Song.Entities/Student_Collect.cs
@@ -22,7 +22,7 @@ namespace Song.Entities {
     		
     		protected String _Qus_Title;
     		
-    		protected Int32 _Sbj_ID;
+    		protected Int64 _Sbj_ID;
     		
     		protected Int32 _Stc_Level;
     		
@@ -100,7 +100,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Sbj_ID {
+    		public Int64 Sbj_ID {
     			get {
     				return this._Sbj_ID;
     			}
@@ -224,7 +224,7 @@ namespace Song.Entities {
     				this._Qus_Title = reader.GetString(_.Qus_Title);
     			}
     			if ((false == reader.IsDBNull(_.Sbj_ID))) {
-    				this._Sbj_ID = reader.GetInt32(_.Sbj_ID);
+    				this._Sbj_ID = reader.GetInt64(_.Sbj_ID);
     			}
     			if ((false == reader.IsDBNull(_.Stc_Level))) {
     				this._Stc_Level = reader.GetInt32(_.Stc_Level);
@@ -297,7 +297,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Qus_Title = new WeiSha.Data.Field<Student_Collect>("Qus_Title");
     			
     			/// <summary>
-    			/// 字段名：Sbj_ID - 数据类型：Int32
+    			/// 字段名：Sbj_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Sbj_ID = new WeiSha.Data.Field<Student_Collect>("Sbj_ID");
