+++ Sourcecode/Song.Entities/Student_Ques.cs
@@ -20,7 +20,7 @@ namespace Song.Entities {
     		
     		protected Int32 _Qus_Diff;
     		
-    		protected Int32 _Sbj_ID;
+    		protected Int64 _Sbj_ID;
     		
     		protected Int32 _Squs_Level;
     		
@@ -86,7 +86,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Sbj_ID {
+    		public Int64 Sbj_ID {
     			get {
     				return this._Sbj_ID;
     			}
@@ -193,7 +193,7 @@ namespace Song.Entities {
     				this._Qus_Diff = reader.GetInt32(_.Qus_Diff);
     			}
     			if ((false == reader.IsDBNull(_.Sbj_ID))) {
-    				this._Sbj_ID = reader.GetInt32(_.Sbj_ID);
+    				this._Sbj_ID = reader.GetInt64(_.Sbj_ID);
     			}
     			if ((false == reader.IsDBNull(_.Squs_Level))) {
     				this._Squs_Level = reader.GetInt32(_.Squs_Level);
@@ -258,7 +258,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Qus_Diff = new WeiSha.Data.Field<Student_Ques>("Qus_Diff");
     			
     			/// <summary>
-    			/// 字段名：Sbj_ID - 数据类型：Int32
+    			/// 字段名：Sbj_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Sbj_ID = new WeiSha.Data.Field<Student_Ques>("Sbj_ID");
