+++ Sourcecode/Song.Entities/TrPlan.cs
@@ -20,7 +20,7 @@ namespace Song.Entities {
     		
     		protected String _Dep_CnName;
     		
-    		protected Int32 _Sbj_ID;
+    		protected Int64 _Sbj_ID;
     		
     		protected String _Sbj_Name;
     		
@@ -104,7 +104,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Sbj_ID {
+    		public Int64 Sbj_ID {
     			get {
     				return this._Sbj_ID;
     			}
@@ -319,7 +319,7 @@ namespace Song.Entities {
     				this._Dep_CnName = reader.GetString(_.Dep_CnName);
     			}
     			if ((false == reader.IsDBNull(_.Sbj_ID))) {
-    				this._Sbj_ID = reader.GetInt32(_.Sbj_ID);
+    				this._Sbj_ID = reader.GetInt64(_.Sbj_ID);
     			}
     			if ((false == reader.IsDBNull(_.Sbj_Name))) {
     				this._Sbj_Name = reader.GetString(_.Sbj_Name);
@@ -411,7 +411,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Dep_CnName = new WeiSha.Data.Field<TrPlan>("Dep_CnName");
     			
     			/// <summary>
-    			/// 字段名：Sbj_ID - 数据类型：Int32
+    			/// 字段名：Sbj_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Sbj_ID = new WeiSha.Data.Field<TrPlan>("Sbj_ID");
